requirements.txt
Telethon
aiohttp
python-dotenv
psutil
python-dateutil
pillow
pytesseract
pyfiglet
ipwhois
emoji
python-whois
requests
deep-translator